refactor(core): move the agent run engine onto the Session
The stream engine — processStream, processStreamChunk, the per-run
StreamState helpers, the subscribed-thread run loop, and the
finish/error handlers — moves out of Harness into a per-session
SessionRunEngine owned by the Session. The engine reaches the host only
through the injected SessionMachinery (resolve the agent, build
run/stream options + toolsets + request context, persist usage, generate
ids, drive tool approve/decline/resume, drain follow-ups); it reads and
writes all per-run state on its own Session.

The pure chunk-to-message content transforms (signal/notification/state
content mappers, usage parsing, finish-reason + server-side-fallback
descriptions) move to a shared stream-content module imported by both
the engine and the Harness signal paths.

resolveToolApproval and the suspension-clearing abort() move onto Session
(pure session state + the already-injected category resolver).

In the multi-user host the run loop, run state, and thread stream are
per-session and cannot be shared, so they belong on the Session; how a
run is produced is shared machinery the Harness owns and injects.

Behavior-preserving: harness.session.processStream(...) and
session.resolveToolApproval(...) replace the previously Harness-private
equivalents. Verified with core + mastracode typecheck, 32 harness test
files, and the full 101-test e2e suite.
