+++ packages/core/src/harness/harness.ts
@@ -6,7 +6,6 @@ import { createSignal, mastraDBMessageToSignal } from '../agent/signals';
 import type { AgentSignalAttributes, AgentSignalContents, AgentSignalInput } from '../agent/signals';
 import type {
   AgentInstructions,
-  AgentThreadSubscription,
   SendAgentNotificationSignalOptions,
   SendAgentNotificationSignalResult,
   ToolsInput,
@@ -14,7 +13,6 @@ import type {
 } from '../agent/types';
 import type { MastraBrowser } from '../browser/browser';
 import { getErrorFromUnknown } from '../error';
-import { getServerSideFallbackInfo } from '../llm/model/server-side-fallback';
 import { Mastra } from '../mastra';
 import type { MastraMemory } from '../memory/memory';
 import type { StorageThreadType } from '../memory/types';
@@ -23,13 +21,22 @@ import type { TracingContext, TracingOptions } from '../observability';
 import { RequestContext } from '../request-context';
 import type { MemoryStorage } from '../storage/domains/memory/base';
 import type { ObservationalMemoryRecord } from '../storage/types';
-import { getTransformedToolPayload, hasTransformedToolPayload } from '../tools/payload-transform';
-import type { ToolPayloadTransformPhase } from '../tools/types';
 import { Workspace } from '../workspace/workspace';
 import type { WorkspaceConfig } from '../workspace/workspace';
 
 import { Session, SessionStream } from './session';
 import type { ThreadDataStore } from './session';
+import {
+  getRecordValue,
+  signalContentsToHarnessContent,
+  signalContentsToText,
+  toNotificationContent,
+  toNotificationSummaryContent,
+  toReactiveSignalContent,
+  toStateSignalContent,
+  toSystemReminderContent,
+  toUserSignalMessage,
+} from './stream-content';
 import {
   askUserTool,
   createSubagentTool,
@@ -51,24 +58,9 @@ import type {
   HarnessSession,
   HarnessThread,
   ModelAuthStatus,
-  PermissionPolicy,
   TokenUsage,
   ToolCategory,
 } from './types';
-
-type HarnessStreamState = {
-  currentMessage: HarnessMessage;
-  lastFinishedMessage?: HarnessMessage;
-  isSuspended: boolean;
-  textContentById: Map<string, { index: number; text: string }>;
-  thinkingContentById: Map<string, { index: number; text: string }>;
-  /**
-   * Set when a stream ends on a non-success finish reason (e.g. `content-filter`,
-   * `error`, `length`). Carries the user-facing message so the run finalizes
-   * into an explicit terminal error state instead of silently completing.
-   */
-  terminalError?: string;
-};
 
 function validateModes(modes: HarnessMode[]): void {
   const modeIds = new Set<string>();
@@ -116,26 +108,6 @@ type HarnessSendNotificationSignalOptions = {
  * silently stop. Returning a message here lets the harness finalize the run
  * into an explicit terminal error state.
  */
-export function describeNonSuccessFinishReason(reason: string, providerMetadata: unknown): string | undefined {
-  switch (reason) {
-    case 'content-filter': {
-      const stopDetails = (providerMetadata as { anthropic?: { stopDetails?: Record<string, unknown> } } | undefined)
-        ?.anthropic?.stopDetails;
-      const explanation =
-        stopDetails && typeof stopDetails.explanation === 'string' ? stopDetails.explanation : undefined;
-      const category = stopDetails && typeof stopDetails.category === 'string' ? stopDetails.category : undefined;
-      const detail = explanation ?? (category ? `category: ${category}` : undefined);
-      return detail ? `The model stopped on a content filter (${detail}).` : 'The model stopped on a content filter.';
-    }
-    case 'error':
-      return 'The model stream ended with an error before producing a final response.';
-    case 'length':
-      return 'The model stopped because it reached its maximum output length before finishing.';
-    default:
-      return undefined;
-  }
-}
-
 /**
  * The Anthropic model that `claude-fable-5` runs are automatically retried on
  * server-side when fable-5's safety classifiers block a turn. See
@@ -189,227 +161,6 @@ export function buildFableFallbackProviderOptions(
  * Without a notice the user has no way to tell that the response did not come
  * from the model they selected.
  */
-export function describeServerSideFallback(providerMetadata: unknown): string | undefined {
-  const fallback = getServerSideFallbackInfo(providerMetadata);
-  if (!fallback) {
-    return undefined;
-  }
-  return fallback.model
-    ? `The selected model declined this turn; the response was generated by fallback model ${fallback.model}.`
-    : 'The selected model declined this turn; the response was generated by a fallback model.';
-}
-
-function getUsageNumber(usage: Record<string, unknown>, key: string): number | undefined {
-  const value = usage[key];
-  if (typeof value === 'number' && Number.isFinite(value)) {
-    return value;
-  }
-  if (typeof value === 'string' && value.trim() !== '') {
-    const numericValue = Number(value);
-    if (Number.isFinite(numericValue)) {
-      return numericValue;
-    }
-  }
-  return undefined;
-}
-
-function addOptionalUsageField(
-  usage: TokenUsage,
-  key: keyof Pick<TokenUsage, 'reasoningTokens' | 'cachedInputTokens' | 'cacheCreationInputTokens'>,
-  value: number | undefined,
-): void {
-  if (value !== undefined) {
-    usage[key] = (usage[key] ?? 0) + value;
-  }
-}
-
-function getDisplayTransform(metadata: unknown, phase: ToolPayloadTransformPhase, fallback: unknown) {
-  const transform = getTransformedToolPayload(metadata, 'display', phase);
-  return hasTransformedToolPayload(transform) ? transform.transformed : fallback;
-}
-
-function getStringValue(value: unknown): string | undefined {
-  return typeof value === 'string' ? value : undefined;
-}
-
-function getRecordValue(value: unknown): Record<string, unknown> | undefined {
-  return value && typeof value === 'object' && !Array.isArray(value) ? (value as Record<string, unknown>) : undefined;
-}
-
-function signalContentsToHarnessContent(contents: AgentSignalContents): HarnessMessageContent[] {
-  if (typeof contents === 'string') return [{ type: 'text', text: contents }];
-  return contents.flatMap((part): HarnessMessageContent[] => {
-    if (part.type === 'text') {
-      return [{ type: 'text', text: part.text }];
-    }
-    if (typeof part.data !== 'string') return [];
-    if (part.mediaType.startsWith('image/')) {
-      return [{ type: 'image', data: part.data, mimeType: part.mediaType }];
-    }
-    return [
-      {
-        type: 'file',
-        data: part.data,
-        mediaType: part.mediaType,
-        filename: part.filename,
-      },
-    ];
-  });
-}
-
-function toSystemReminderContent(
-  payload: Record<string, unknown>,
-): Extract<HarnessMessageContent, { type: 'system_reminder' }> | undefined {
-  const attributes = getRecordValue(payload.attributes);
-  const metadata = getRecordValue(payload.metadata);
-  const message = signalContentsToText(payload.contents);
-  if (!message) return undefined;
-
-  return {
-    type: 'system_reminder',
-    message,
-    reminderType:
-      getStringValue(payload.reminderType) ?? getStringValue(attributes?.type) ?? getStringValue(payload.type),
-    path: getStringValue(payload.path) ?? getStringValue(attributes?.path),
-    precedesMessageId: getStringValue(payload.precedesMessageId) ?? getStringValue(attributes?.precedesMessageId),
-    gapText: getStringValue(payload.gapText) ?? getStringValue(attributes?.gapText),
-    gapMs:
-      typeof payload.gapMs === 'number'
-        ? payload.gapMs
-        : typeof attributes?.gapMs === 'number'
-          ? attributes.gapMs
-          : undefined,
-    timestamp: getStringValue(payload.timestamp) ?? getStringValue(attributes?.timestamp),
-    goalMaxTurns:
-      typeof payload.goalMaxTurns === 'number'
-        ? payload.goalMaxTurns
-        : typeof metadata?.goalMaxTurns === 'number'
-          ? metadata.goalMaxTurns
-          : undefined,
-    judgeModelId: getStringValue(payload.judgeModelId) ?? getStringValue(metadata?.judgeModelId),
-    goalEvaluation: getRecordValue(metadata?.goalEvaluation) as
-      | Extract<HarnessMessageContent, { type: 'system_reminder' }>['goalEvaluation']
-      | undefined,
-  };
-}
-
-function toUserSignalMessage(payload: Record<string, unknown>): HarnessMessage | undefined {
-  const id = getStringValue(payload.id);
-  const rawContents = payload.contents;
-  if (!id || rawContents === undefined) return undefined;
-
-  const signal = createSignal({
-    id,
-    type: 'user',
-    tagName: 'user',
-    contents: rawContents as AgentSignalContents,
-    attributes: getRecordValue(payload.attributes) as AgentSignalInput['attributes'],
-    createdAt: getStringValue(payload.createdAt),
-  });
-  const content = signalContentsToHarnessContent(signal.contents);
-  if (content.length === 0) return undefined;
-
-  return {
-    id: signal.id,
-    role: 'user',
-    content,
-    createdAt: signal.createdAt,
-    attributes: signal.attributes,
-  };
-}
-
-function signalContentsToText(contents: unknown): string {
-  if (typeof contents === 'string') return contents;
-  if (!Array.isArray(contents)) return '';
-  return contents
-    .filter((part): part is { type: 'text'; text: string } => getRecordValue(part)?.type === 'text')
-    .map(part => part.text)
-    .join('\n');
-}
-
-function toStateSignalContent(
-  payload: Record<string, unknown>,
-): Extract<HarnessMessageContent, { type: 'state_signal' }> | undefined {
-  const stateMetadata = getRecordValue(getRecordValue(payload.metadata)?.state);
-  const stateId = getStringValue(stateMetadata?.id) ?? getStringValue(payload.tagName) ?? 'state';
-
-  return {
-    type: 'state_signal',
-    id: getStringValue(payload.id),
-    stateId,
-    mode: stateMetadata?.mode === 'delta' ? 'delta' : 'snapshot',
-    cacheKey: getStringValue(stateMetadata?.cacheKey),
-    version: typeof stateMetadata?.version === 'number' ? stateMetadata.version : undefined,
-    message: signalContentsToText(payload.contents),
-  };
-}
-
-function toNotificationSummaryContent(
-  payload: Record<string, unknown>,
-): Extract<HarnessMessageContent, { type: 'notification_summary' }> | undefined {
-  const metadataSummary = getRecordValue(getRecordValue(payload.metadata)?.notificationSummary);
-  const bySource = getRecordValue(metadataSummary?.bySource) ?? {};
-  const byPriority = getRecordValue(metadataSummary?.byPriority) ?? {};
-  const notificationIds = Array.isArray(metadataSummary?.notificationIds)
-    ? metadataSummary.notificationIds.filter((id): id is string => typeof id === 'string')
-    : [];
-  const pending = typeof metadataSummary?.pending === 'number' ? metadataSummary.pending : undefined;
-
-  return {
-    type: 'notification_summary',
-    id: getStringValue(payload.id),
-    message: signalContentsToText(payload.contents),
-    pending: pending ?? notificationIds.length,
-    bySource: Object.fromEntries(
-      Object.entries(bySource).filter((entry): entry is [string, number] => typeof entry[1] === 'number'),
-    ),
-    byPriority: Object.fromEntries(
-      Object.entries(byPriority).filter((entry): entry is [string, number] => typeof entry[1] === 'number'),
-    ),
-    notificationIds,
-  };
-}
-
-function toReactiveSignalContent(
-  payload: Record<string, unknown>,
-): Extract<HarnessMessageContent, { type: 'reactive_signal' }> | undefined {
-  const tagName = getStringValue(payload.tagName);
-  if (!tagName) return undefined;
-
-  return {
-    type: 'reactive_signal',
-    id: getStringValue(payload.id),
-    tagName,
-    message: signalContentsToText(payload.contents),
-    attributes: getRecordValue(payload.attributes),
-    metadata: getRecordValue(payload.metadata),
-  };
-}
-
-function toNotificationContent(
-  payload: Record<string, unknown>,
-): Extract<HarnessMessageContent, { type: 'notification' }> | undefined {
-  const attributes = getRecordValue(payload.attributes) ?? {};
-  const metadata = getRecordValue(payload.metadata) ?? {};
-  const notificationMetadata = getRecordValue(metadata.notification);
-  const message = signalContentsToText(payload.contents);
-  if (!message) return undefined;
-
-  return {
-    type: 'notification',
-    id: getStringValue(payload.id),
-    notificationId: getStringValue(attributes.id) ?? getStringValue(notificationMetadata?.recordId),
-    message,
-    source: getStringValue(attributes.source) ?? getStringValue(notificationMetadata?.source),
-    kind:
-      getStringValue(attributes.kind) ?? getStringValue(attributes.type) ?? getStringValue(notificationMetadata?.kind),
-    priority: getStringValue(attributes.priority) ?? getStringValue(notificationMetadata?.priority),
-    status: getStringValue(attributes.status) ?? getStringValue(notificationMetadata?.status),
-    attributes,
-    metadata,
-  };
-}
-
 /**
  * The Harness orchestrates multiple agent modes, shared state, memory, and storage.
  * It's the core abstraction that a TUI (or other UI) controls.
@@ -556,6 +307,12 @@ export class Harness<TState = {}> {
       buildRequestContext: requestContext => this.buildRequestContext(requestContext),
       persistTokenUsage: () => this.persistTokenUsage(),
       generateId: () => this.generateId(),
+      approveToolCall: input => this.handleToolApprove(input),
+      declineToolCall: input => this.handleToolDecline(input),
+      resumeToolCall: input => this.handleToolResume(input),
+      drainFollowUpQueue: async () => {
+        await this.drainFollowUpQueue();
+      },
     });
 
     // Seed the selected model: an explicit initialState.currentModelId wins,
@@ -1615,34 +1372,6 @@ export class Harness<TState = {}> {
     return this.config.toolCategoryResolver?.(toolName) ?? null;
   }
 
-  /**
-   * Resolve whether a tool call should be auto-approved, denied, or asked.
-   * Resolution chain: per-tool deny → yolo → per-tool policy → session tool grant →
-   * session category grant → category policy → "ask"
-   */
-  private resolveToolApproval(toolName: string): PermissionPolicy {
-    const state = this.#session.state.get() as Record<string, unknown>;
-    const rules = this.#session.permissions.getRules();
-
-    const toolPolicy = rules.tools[toolName];
-    if (toolPolicy === 'deny') return 'deny';
-
-    if (state.yolo === true) return 'allow';
-
-    if (toolPolicy) return toolPolicy;
-
-    if (this.#session.hasToolGrant(toolName)) return 'allow';
-
-    const category = this.getToolCategory({ toolName });
-    if (category) {
-      if (this.#session.hasCategoryGrant(category)) return 'allow';
-      const categoryPolicy = rules.categories[category];
-      if (categoryPolicy) return categoryPolicy;
-    }
-
-    return 'ask';
-  }
-
   // ===========================================================================
   // Message Handling
   // ===========================================================================
@@ -1662,7 +1391,7 @@ export class Harness<TState = {}> {
       threadId,
     });
     this.#session.stream.attach({ subscription, key });
-    void this.processSubscribedThreadStream(subscription);
+    void this.#session.processSubscribedThreadStream(subscription);
   }
 
   private async ensureCurrentAgentThreadSubscription(): Promise<void> {
@@ -1816,118 +1545,6 @@ export class Harness<TState = {}> {
       this.#session.followUps.requeue(next);
       this.#session.emit({ type: 'follow_up_queued', count: this.#session.followUps.count() });
       throw error;
-    }
-  }
-
-  private async finishSubscribedStreamRun({
-    suspended,
-    error,
-    aborted,
-  }: {
-    suspended?: boolean;
-    error?: boolean;
-    aborted?: boolean;
-  }): Promise<void> {
-    const reason = error
-      ? 'error'
-      : suspended
-        ? 'suspended'
-        : aborted || this.#session.run.isAbortRequested()
-          ? 'aborted'
-          : 'complete';
-    this.#session.emit({ type: 'agent_end', reason });
-    this.#session.run.reset();
-    await this.drainFollowUpQueue();
-  }
-
-  private async handleSubscribedStreamError(error: unknown): Promise<void> {
-    if (error instanceof Error && error.name === 'AbortError') {
-      this.#session.emit({ type: 'agent_end', reason: 'aborted' });
-    } else {
-      this.#session.emit({ type: 'error', error: getErrorFromUnknown(error) });
-      this.#session.emit({ type: 'agent_end', reason: 'error' });
-    }
-    this.#session.stream.detach();
-    this.#session.run.reset();
-    await this.drainFollowUpQueue();
-  }
-
-  private async processSubscribedThreadStream(subscription: AgentThreadSubscription<any>): Promise<void> {
-    const requestContext = await this.buildRequestContext();
-    let currentRun: HarnessStreamState | undefined;
-    let lastFinishedRunId: string | null = null;
-
-    try {
-      for await (const chunk of subscription.stream) {
-        if (!this.#session.stream.isCurrent({ subscription })) {
-          subscription.unsubscribe();
-          break;
-        }
-
-        const chunkRunId = 'runId' in chunk ? chunk.runId : null;
-        if (lastFinishedRunId && chunkRunId === lastFinishedRunId) {
-          continue;
-        }
-
-        if (!currentRun) {
-          currentRun = this.createStreamState();
-          this.#session.run.nextOperation();
-          this.#session.run.ensureAbortController();
-          this.#session.run.setRunId({ runId: subscription.activeRunId() ?? ('runId' in chunk ? chunk.runId : null) });
-          this.#session.run.setTraceId({ traceId: null });
-          this.#session.emit({ type: 'agent_start' });
-        }
-
-        if (chunk.type === 'start') {
-          continue;
-        }
-
-        try {
-          const streamResult = await this.processStreamChunk(currentRun, chunk, requestContext);
-          if (
-            streamResult ||
-            chunk.type === 'finish' ||
-            chunk.type === 'error' ||
-            chunk.type === 'abort' ||
-            chunk.type === 'tool-call-suspended'
-          ) {
-            const finishedRunId: string | null = chunkRunId ?? this.#session.run.getRunId();
-            const suspended =
-              chunk.type === 'tool-call-suspended' ||
-              (streamResult ?? this.finishStreamState(currentRun)).suspended ||
-              undefined;
-            const aborted = chunk.type === 'abort';
-            // A non-success terminal finish reason (e.g. a `claude-fable-5`
-            // content-filter refusal) is surfaced as an explicit error so the
-            // run never silently stops without a visible terminal state.
-            let isError = chunk.type === 'error';
-            if (
-              currentRun.terminalError &&
-              !isError &&
-              !aborted &&
-              !this.#session.run.isAbortRequested() &&
-              !suspended
-            ) {
-              isError = true;
-              this.#session.emit({ type: 'error', error: new Error(currentRun.terminalError) });
-            }
-            await this.finishSubscribedStreamRun({
-              suspended,
-              error: isError,
-              aborted,
-            });
-            lastFinishedRunId = finishedRunId;
-            currentRun = undefined;
-          }
-        } catch (error) {
-          await this.handleSubscribedStreamError(error);
-          currentRun = undefined;
-        }
-      }
-    } catch (error) {
-      if (this.#session.stream.isCurrent({ subscription })) {
-        await this.handleSubscribedStreamError(error);
-      }
     }
   }
 
@@ -2446,680 +2063,6 @@ export class Harness<TState = {}> {
     return { id: msg.id, role: msg.role === 'signal' ? 'user' : msg.role, content, createdAt: msg.createdAt };
   }
 
-  private createEmptyAssistantMessage(): HarnessMessage {
-    return {
-      id: this.generateId(),
-      role: 'assistant',
-      content: [],
-      createdAt: new Date(),
-    };
-  }
-
-  private hasCurrentMessageContent(state: HarnessStreamState): boolean {
-    return state.currentMessage.content.length > 0 || Boolean(state.currentMessage.stopReason);
-  }
-
-  private finishCurrentMessageAndRotate(state: HarnessStreamState): void {
-    if (!this.hasCurrentMessageContent(state)) return;
-    this.#session.emit({ type: 'message_end', message: state.currentMessage });
-    state.lastFinishedMessage = state.currentMessage;
-    state.currentMessage = this.createEmptyAssistantMessage();
-    state.textContentById.clear();
-    state.thinkingContentById.clear();
-  }
-
-  /**
-   * Process a stream response (shared between sendMessage and tool approval).
-   */
-  private createStreamState(): HarnessStreamState {
-    return {
-      currentMessage: this.createEmptyAssistantMessage(),
-      isSuspended: false,
-      textContentById: new Map<string, { index: number; text: string }>(),
-      thinkingContentById: new Map<string, { index: number; text: string }>(),
-    };
-  }
-
-  private abortForOmFailure({ operationType, stage, error }: { operationType: string; stage: string; error: string }) {
-    this.#session.emit({
-      type: 'error',
-      error: new Error(`Observational memory ${operationType} ${stage} failed: ${error}`),
-    });
-    this.abort();
-  }
-
-  private async processStream(
-    response: { fullStream: AsyncIterable<any> },
-    requestContextInput?: RequestContext,
-  ): Promise<{ message: HarnessMessage; suspended?: boolean } | undefined> {
-    const state = this.createStreamState();
-    const requestContext = await this.buildRequestContext(requestContextInput);
-    this.#session.run.nextOperation();
-    this.#session.emit({ type: 'agent_start' });
-
-    let result: { message: HarnessMessage; suspended?: boolean } | undefined;
-    let error = false;
-    let aborted = false;
-
-    for await (const chunk of response.fullStream) {
-      result = await this.processStreamChunk(state, chunk, requestContext);
-      if (chunk.type === 'error') {
-        error = true;
-      }
-      if (chunk.type === 'abort') {
-        aborted = true;
-      }
-      if (
-        result ||
-        chunk.type === 'finish' ||
-        chunk.type === 'error' ||
-        chunk.type === 'abort' ||
-        chunk.type === 'tool-call-suspended' ||
-        this.#session.run.isAbortRequested()
-      ) {
-        result ??= this.finishStreamState(state);
-        break;
-      }
-    }
-
-    result ??= this.finishStreamState(state);
-
-    // A non-success terminal finish reason (e.g. a `claude-fable-5`
-    // content-filter refusal) is surfaced as an explicit error so the run never
-    // silently stops without a visible terminal state.
-    if (state.terminalError && !error && !aborted && !this.#session.run.isAbortRequested() && !result.suspended) {
-      error = true;
-      this.#session.emit({ type: 'error', error: new Error(state.terminalError) });
-    }
-
-    this.#session.emit({
-      type: 'agent_end',
-      reason: error
-        ? 'error'
-        : result.suspended
-          ? 'suspended'
-          : aborted || this.#session.run.isAbortRequested()
-            ? 'aborted'
-            : 'complete',
-    });
-
-    this.#session.run.reset();
-    await this.drainFollowUpQueue();
-
-    return result;
-  }
-
-  private async processStreamChunk(
-    state: HarnessStreamState,
-    chunk: any,
-    requestContext: RequestContext,
-  ): Promise<{ message: HarnessMessage; suspended?: boolean } | undefined> {
-    if ('runId' in chunk && chunk.runId) {
-      this.#session.run.setRunId({ runId: chunk.runId });
-    }
-
-    switch (chunk.type) {
-      case 'text-start': {
-        const textIndex = state.currentMessage.content.length;
-        state.currentMessage.content.push({ type: 'text', text: '' });
-        state.textContentById.set(chunk.payload.id, { index: textIndex, text: '' });
-        this.#session.emit({ type: 'message_start', message: { ...state.currentMessage } });
-        break;
-      }
-
-      case 'text-delta': {
-        const textState = state.textContentById.get(chunk.payload.id);
-        if (textState) {
-          textState.text += chunk.payload.text;
-          const textContent = state.currentMessage.content[textState.index];
-          if (textContent && textContent.type === 'text') {
-            textContent.text = textState.text;
-          }
-          this.#session.emit({ type: 'message_update', message: { ...state.currentMessage } });
-        }
-        break;
-      }
-
-      case 'reasoning-start': {
-        const thinkingIndex = state.currentMessage.content.length;
-        state.currentMessage.content.push({ type: 'thinking', thinking: '' });
-        state.thinkingContentById.set(chunk.payload.id, { index: thinkingIndex, text: '' });
-        this.#session.emit({ type: 'message_update', message: { ...state.currentMessage } });
-        break;
-      }
-
-      case 'reasoning-delta': {
-        const thinkingState = state.thinkingContentById.get(chunk.payload.id);
-        if (thinkingState) {
-          thinkingState.text += chunk.payload.text;
-          const thinkingContent = state.currentMessage.content[thinkingState.index];
-          if (thinkingContent && thinkingContent.type === 'thinking') {
-            thinkingContent.thinking = thinkingState.text;
-          }
-          this.#session.emit({ type: 'message_update', message: { ...state.currentMessage } });
-        }
-        break;
-      }
-
-      case 'tool-call-input-streaming-start': {
-        const { toolCallId, toolName } = chunk.payload;
-        this.#session.emit({ type: 'tool_input_start', toolCallId, toolName });
-        break;
-      }
-
-      case 'tool-call-delta': {
-        const { toolCallId, argsTextDelta, toolName } = chunk.payload;
-        const transform = getTransformedToolPayload(chunk.metadata, 'display', 'input-delta');
-        if (!transform?.suppress) {
-          this.#session.emit({
-            type: 'tool_input_delta',
-            toolCallId,
-            argsTextDelta: hasTransformedToolPayload(transform) ? transform.transformed : argsTextDelta,
-            toolName,
-          });
-        }
-        break;
-      }
-
-      case 'tool-call-input-streaming-end': {
-        const { toolCallId } = chunk.payload;
-        this.#session.emit({ type: 'tool_input_end', toolCallId });
-        break;
-      }
-
-      case 'tool-call': {
-        const toolCall = chunk.payload;
-        const args = getDisplayTransform(chunk.metadata, 'input-available', toolCall.args);
-        state.currentMessage.content.push({
-          type: 'tool_call',
-          id: toolCall.toolCallId,
-          name: toolCall.toolName,
-          args,
-        });
-        this.#session.emit({
-          type: 'tool_start',
-          toolCallId: toolCall.toolCallId,
-          toolName: toolCall.toolName,
-          args,
-        });
-        this.#session.emit({ type: 'message_update', message: { ...state.currentMessage } });
-        break;
-      }
-
-      case 'tool-result': {
-        const toolResult = chunk.payload;
-        const providerMetadata = toolResult.providerMetadata as Record<string, unknown> | undefined;
-        const result = getDisplayTransform(chunk.metadata, 'output-available', toolResult.result);
-        state.currentMessage.content.push({
-          type: 'tool_result',
-          id: toolResult.toolCallId,
-          name: toolResult.toolName,
-          result,
-          isError: toolResult.isError ?? false,
-          ...(providerMetadata ? { providerMetadata } : {}),
-        });
-        this.#session.emit({
-          type: 'tool_end',
-          toolCallId: toolResult.toolCallId,
-          result,
-          isError: toolResult.isError ?? false,
-          ...(providerMetadata ? { providerMetadata } : {}),
-        });
-        this.#session.emit({ type: 'message_update', message: { ...state.currentMessage } });
-        break;
-      }
-
-      case 'tool-error': {
-        const toolError = chunk.payload;
-        this.#session.emit({
-          type: 'tool_end',
-          toolCallId: toolError.toolCallId,
-          result: getDisplayTransform(chunk.metadata, 'error', toolError.error),
-          isError: true,
-        });
-        break;
-      }
-
-      case 'tool-call-approval': {
-        const toolCallId = chunk.payload.toolCallId;
-        const toolName = chunk.payload.toolName;
-        const approvalTransform = getTransformedToolPayload(chunk.metadata, 'display', 'approval');
-        const toolArgs = hasTransformedToolPayload(approvalTransform)
-          ? approvalTransform.transformed
-          : getDisplayTransform(chunk.metadata, 'input-available', chunk.payload.args);
-
-        const policy = this.resolveToolApproval(toolName);
-
-        if (policy === 'allow') {
-          await this.handleToolApprove({ toolCallId, requestContext });
-          break;
-        }
-
-        if (policy === 'deny') {
-          await this.handleToolDecline({ toolCallId, requestContext });
-          break;
-        }
-
-        const approvalPromise = this.#session.approval.arm({ toolName });
-        this.#session.emit({ type: 'tool_approval_required', toolCallId, toolName, args: toolArgs });
-
-        const approval = await approvalPromise;
-        this.#session.approval.clearToolName();
-
-        if (approval.decision === 'approve') {
-          await this.handleToolApprove({ toolCallId, requestContext: approval.requestContext ?? requestContext });
-        } else {
-          await this.handleToolDecline({ toolCallId, requestContext: approval.requestContext ?? requestContext });
-        }
-        break;
-      }
-
-      case 'tool-call-suspended': {
-        const suspToolCallId = chunk.payload.toolCallId;
-        const suspToolName = chunk.payload.toolName;
-        const suspArgs = getDisplayTransform(chunk.metadata, 'input-available', chunk.payload.args);
-        const suspPayload = getDisplayTransform(chunk.metadata, 'suspend', chunk.payload.suspendPayload);
-        const suspResumeSchema = chunk.payload.resumeSchema;
-
-        const suspRunId = this.#session.run.getRunId();
-        if (suspRunId) {
-          this.#session.suspensions.register({
-            toolCallId: suspToolCallId,
-            runId: suspRunId,
-            toolName: suspToolName,
-          });
-        }
-        state.isSuspended = true;
-
-        this.#session.emit({
-          type: 'tool_suspended',
-          toolCallId: suspToolCallId,
-          toolName: suspToolName,
-          args: suspArgs,
-          suspendPayload: suspPayload,
-          resumeSchema: suspResumeSchema,
-        });
-
-        break;
-      }
-
-      case 'error': {
-        const streamError = getErrorFromUnknown(chunk.payload.error);
-        this.#session.emit({ type: 'error', error: streamError });
-        break;
-      }
-
-      case 'step-finish': {
-        const usage = chunk.payload?.output?.usage;
-        if (usage) {
-          const usageRecord = usage as Record<string, unknown>;
-          const promptTokens =
-            getUsageNumber(usageRecord, 'promptTokens') ?? getUsageNumber(usageRecord, 'inputTokens') ?? 0;
-          const completionTokens =
-            getUsageNumber(usageRecord, 'completionTokens') ?? getUsageNumber(usageRecord, 'outputTokens') ?? 0;
-          const totalTokens = getUsageNumber(usageRecord, 'totalTokens') ?? promptTokens + completionTokens;
-          const stepUsage: TokenUsage = {
-            promptTokens,
-            completionTokens,
-            totalTokens,
-          };
-          addOptionalUsageField(stepUsage, 'reasoningTokens', getUsageNumber(usageRecord, 'reasoningTokens'));
-          addOptionalUsageField(stepUsage, 'cachedInputTokens', getUsageNumber(usageRecord, 'cachedInputTokens'));
-          addOptionalUsageField(
-            stepUsage,
-            'cacheCreationInputTokens',
-            getUsageNumber(usageRecord, 'cacheCreationInputTokens'),
-          );
-          if (usageRecord.raw !== undefined) {
-            stepUsage.raw = usageRecord.raw;
-          }
-
-          this.#session.addUsage(stepUsage);
-
-          this.persistTokenUsage().catch(() => {});
-          this.#session.emit({ type: 'usage_update', usage: stepUsage });
-        }
-        break;
-      }
-
-      case 'finish': {
-        const finishReason = chunk.payload.stepResult?.reason;
-        const finishProviderMetadata = chunk.payload?.metadata?.providerMetadata ?? chunk.payload?.providerMetadata;
-        // A server-side fallback means the turn was answered by a different
-        // model than the one the user selected (e.g. fable-5 declined and the
-        // fallback served the response). Surface that, otherwise the
-        // substitution is invisible.
-        const fallbackNotice = describeServerSideFallback(finishProviderMetadata);
-        if (fallbackNotice) {
-          this.#session.emit({ type: 'info', message: fallbackNotice });
-        }
-        if (finishReason === 'stop' || finishReason === 'end-turn') {
-          state.currentMessage.stopReason = 'complete';
-        } else if (finishReason === 'tool-calls') {
-          state.currentMessage.stopReason = 'tool_use';
-        } else {
-          // Non-success terminal reasons (e.g. `content-filter` from a
-          // `claude-fable-5` refusal, `error`, or `length`) must surface as an
-          // explicit terminal error rather than a silent `complete`. Otherwise
-          // the run ends with no final message and no error, leaving the user
-          // unable to tell whether it completed, failed, or is still active.
-          const errorMessage = describeNonSuccessFinishReason(finishReason, finishProviderMetadata);
-          if (errorMessage) {
-            state.currentMessage.stopReason = 'error';
-            state.currentMessage.errorMessage = errorMessage;
-            state.terminalError = errorMessage;
-          } else {
-            state.currentMessage.stopReason = 'complete';
-          }
-        }
-        break;
-      }
-
-      case 'goal': {
-        // In-loop goal evaluation marks a boundary between assistant attempts.
-        // Close the current assistant message before rendering the judge result
-        // so a continuation starts a fresh message instead of overwriting the
-        // previous attempt in streaming UIs.
-        this.finishCurrentMessageAndRotate(state);
-        // Forward the payload so consumers (the TUI's judge display) can render
-        // judge progress and the decision.
-        this.#session.emit({ type: 'goal_evaluation', payload: chunk.payload });
-        break;
-      }
-
-      // Observational Memory data parts
-      // NOTE: OM data parts arrive as { type, data: { ... } } — NOT { type, payload }
-      case 'data-om-status': {
-        const d = (chunk as any).data as Record<string, any> | undefined;
-        if (d?.windows) {
-          const w = d.windows;
-          const active = w.active ?? {};
-          const msgs = active.messages ?? {};
-          const obs = active.observations ?? {};
-          const buffObs = w.buffered?.observations ?? {};
-          const buffRef = w.buffered?.reflection ?? {};
-
-          this.#session.emit({
-            type: 'om_status',
-            windows: {
-              active: {
-                messages: { tokens: msgs.tokens ?? 0, threshold: msgs.threshold ?? 0 },
-                observations: { tokens: obs.tokens ?? 0, threshold: obs.threshold ?? 0 },
-              },
-              buffered: {
-                observations: {
-                  status: buffObs.status ?? 'idle',
-                  chunks: buffObs.chunks ?? 0,
-                  messageTokens: buffObs.messageTokens ?? 0,
-                  projectedMessageRemoval: buffObs.projectedMessageRemoval ?? 0,
-                  observationTokens: buffObs.observationTokens ?? 0,
-                },
-                reflection: {
-                  status: buffRef.status ?? 'idle',
-                  inputObservationTokens: buffRef.inputObservationTokens ?? 0,
-                  observationTokens: buffRef.observationTokens ?? 0,
-                },
-              },
-            },
-            recordId: d.recordId ?? '',
-            threadId: d.threadId ?? '',
-            stepNumber: d.stepNumber ?? 0,
-            generationCount: d.generationCount ?? 0,
-          });
-        }
-        break;
-      }
-      case 'data-om-observation-start': {
-        const payload = (chunk as any).data as Record<string, any> | undefined;
-        if (payload && payload.cycleId) {
-          if (payload.operationType === 'observation') {
-            this.#session.emit({
-              type: 'om_observation_start',
-              cycleId: payload.cycleId,
-              operationType: payload.operationType,
-              tokensToObserve: payload.tokensToObserve ?? 0,
-            });
-          } else if (payload.operationType === 'reflection') {
-            this.#session.emit({
-              type: 'om_reflection_start',
-              cycleId: payload.cycleId,
-              tokensToReflect: payload.tokensToObserve ?? 0,
-            });
-          }
-        }
-        break;
-      }
-      case 'data-om-observation-end': {
-        const payload = (chunk as any).data as Record<string, any> | undefined;
-        if (payload && payload.cycleId) {
-          if (payload.operationType === 'reflection') {
-            this.#session.emit({
-              type: 'om_reflection_end',
-              cycleId: payload.cycleId,
-              durationMs: payload.durationMs ?? 0,
-              compressedTokens: payload.observationTokens ?? 0,
-              observations: payload.observations,
-            });
-          } else {
-            this.#session.emit({
-              type: 'om_observation_end',
-              cycleId: payload.cycleId,
-              durationMs: payload.durationMs ?? 0,
-              tokensObserved: payload.tokensObserved ?? 0,
-              observationTokens: payload.observationTokens ?? 0,
-              observations: payload.observations,
-              currentTask: payload.currentTask,
-              suggestedResponse: payload.suggestedResponse,
-            });
-          }
-        }
-        break;
-      }
-      case 'data-om-observation-failed': {
-        const payload = (chunk as any).data as Record<string, any> | undefined;
-        if (payload) {
-          const operationType = payload.operationType === 'reflection' ? 'reflection' : 'observation';
-          const error = payload.error ?? 'Unknown error';
-
-          if (operationType === 'reflection') {
-            this.#session.emit({
-              type: 'om_reflection_failed',
-              cycleId: payload.cycleId ?? 'unknown',
-              error,
-              durationMs: payload.durationMs ?? 0,
-            });
-          } else {
-            this.#session.emit({
-              type: 'om_observation_failed',
-              cycleId: payload.cycleId ?? 'unknown',
-              error,
-              durationMs: payload.durationMs ?? 0,
-            });
-          }
-
-          this.abortForOmFailure({ operationType, stage: 'run', error });
-          return { message: state.currentMessage };
-        }
-        break;
-      }
-      // Async buffering lifecycle
-      case 'data-om-buffering-start': {
-        const payload = (chunk as any).data as Record<string, any> | undefined;
-        if (payload && payload.cycleId) {
-          this.#session.emit({
-            type: 'om_buffering_start',
-            cycleId: payload.cycleId,
-            operationType: payload.operationType ?? 'observation',
-            tokensToBuffer: payload.tokensToBuffer ?? 0,
-          });
-        }
-        break;
-      }
-      case 'data-om-buffering-end': {
-        const payload = (chunk as any).data as Record<string, any> | undefined;
-        if (payload && payload.cycleId) {
-          this.#session.emit({
-            type: 'om_buffering_end',
-            cycleId: payload.cycleId,
-            operationType: payload.operationType ?? 'observation',
-            tokensBuffered: payload.tokensBuffered ?? 0,
-            bufferedTokens: payload.bufferedTokens ?? 0,
-            observations: payload.observations,
-          });
-        }
-        break;
-      }
-      case 'data-om-buffering-failed': {
-        const payload = (chunk as any).data as Record<string, any> | undefined;
-        if (payload) {
-          const operationType = payload.operationType ?? 'observation';
-          const error = payload.error ?? 'Unknown error';
-
-          this.#session.emit({
-            type: 'om_buffering_failed',
-            cycleId: payload.cycleId,
-            operationType,
-            error,
-          });
-
-          this.abortForOmFailure({ operationType, stage: 'buffering', error });
-          return { message: state.currentMessage };
-        }
-        break;
-      }
-      case 'data-signal': {
-        const payload = (chunk as any).data as Record<string, unknown> | undefined;
-        if (payload?.type === 'state') {
-          const stateSignal = toStateSignalContent(payload);
-          if (stateSignal) {
-            state.currentMessage.content.push(stateSignal);
-            this.#session.emit({ type: 'message_update', message: state.currentMessage });
-          }
-        } else if (payload?.type === 'reactive' && payload.tagName === 'system-reminder') {
-          const reminder = toSystemReminderContent(payload);
-          if (reminder) {
-            state.currentMessage.content.push(reminder);
-            this.#session.emit({ type: 'message_update', message: state.currentMessage });
-          }
-        } else if (payload?.type === 'notification' && payload.tagName === 'notification-summary') {
-          const notificationSummary = toNotificationSummaryContent(payload);
-          if (notificationSummary) {
-            state.currentMessage.content.push(notificationSummary);
-            this.#session.emit({ type: 'message_update', message: state.currentMessage });
-          }
-        } else if (payload?.type === 'notification' && payload.tagName === 'notification') {
-          const notification = toNotificationContent(payload);
-          if (notification) {
-            state.currentMessage.content.push(notification);
-            this.#session.emit({ type: 'message_update', message: state.currentMessage });
-          }
-        } else if (payload?.type === 'reactive') {
-          const reactiveSignal = toReactiveSignalContent(payload);
-          if (reactiveSignal) {
-            state.currentMessage.content.push(reactiveSignal);
-            this.#session.emit({ type: 'message_update', message: state.currentMessage });
-          }
-        }
-        break;
-      }
-      case 'data-user-message': {
-        const payload = (chunk as any).data as Record<string, unknown> | undefined;
-        const message = payload ? toUserSignalMessage(payload) : undefined;
-        if (message) {
-          if (state.currentMessage.content.length > 0) {
-            state.currentMessage.stopReason ??= 'complete';
-            this.#session.emit({ type: 'message_end', message: { ...state.currentMessage } });
-            state.currentMessage = {
-              id: this.generateId(),
-              role: 'assistant',
-              content: [],
-              createdAt: new Date(),
-            };
-            state.textContentById.clear();
-            state.thinkingContentById.clear();
-          }
-          this.#session.emit({ type: 'message_start', message });
-          this.#session.emit({ type: 'message_end', message });
-        }
-        break;
-      }
-      // Back-compat: persisted streams may still contain data-system-reminder parts
-      case 'data-system-reminder': {
-        const payload = (chunk as any).data as Record<string, unknown> | undefined;
-        const reminder = payload ? toSystemReminderContent(payload) : undefined;
-        if (reminder) {
-          state.currentMessage.content.push(reminder);
-          this.#session.emit({ type: 'message_update', message: state.currentMessage });
-        }
-        break;
-      }
-      case 'data-om-activation': {
-        const payload = (chunk as any).data as Record<string, any> | undefined;
-        if (payload && payload.cycleId) {
-          this.#session.emit({
-            type: 'om_activation',
-            cycleId: payload.cycleId,
-            operationType: payload.operationType ?? 'observation',
-            chunksActivated: payload.chunksActivated ?? 0,
-            tokensActivated: payload.tokensActivated ?? 0,
-            observationTokens: payload.observationTokens ?? 0,
-            messagesActivated: payload.messagesActivated ?? 0,
-            generationCount: payload.generationCount ?? 0,
-            triggeredBy: payload.triggeredBy,
-            lastActivityAt: payload.lastActivityAt,
-            ttlExpiredMs: payload.ttlExpiredMs,
-            activateAfterIdle:
-              typeof payload.config?.activateAfterIdle === 'number' ? payload.config.activateAfterIdle : undefined,
-            previousModel: payload.previousModel,
-            currentModel: payload.currentModel,
-          });
-        }
-        break;
-      }
-      case 'data-om-thread-update': {
-        const payload = (chunk as any).data as Record<string, any> | undefined;
-        if (payload && payload.newTitle) {
-          this.#session.emit({
-            type: 'om_thread_title_updated',
-            cycleId: payload.cycleId ?? 'unknown',
-            threadId: payload.threadId ?? this.#session.thread.getId() ?? 'unknown',
-            oldTitle: payload.oldTitle,
-            newTitle: payload.newTitle,
-          });
-        }
-        break;
-      }
-
-      // Sandbox streaming data chunks (from workspace execute_command tool)
-      case 'data-sandbox-stdout': {
-        const d = (chunk as any).data as Record<string, any> | undefined;
-        if (d?.output && d?.toolCallId) {
-          this.#session.emit({ type: 'shell_output', toolCallId: d.toolCallId, output: d.output, stream: 'stdout' });
-        }
-        break;
-      }
-      case 'data-sandbox-stderr': {
-        const d = (chunk as any).data as Record<string, any> | undefined;
-        if (d?.output && d?.toolCallId) {
-          this.#session.emit({ type: 'shell_output', toolCallId: d.toolCallId, output: d.output, stream: 'stderr' });
-        }
-        break;
-      }
-
-      default:
-        break;
-    }
-  }
-
-  private finishStreamState(state: HarnessStreamState): { message: HarnessMessage; suspended?: boolean } {
-    if (this.hasCurrentMessageContent(state) || !state.lastFinishedMessage) {
-      this.#session.emit({ type: 'message_end', message: state.currentMessage });
-      return { message: state.currentMessage, suspended: state.isSuspended || undefined };
-    }
-
-    return { message: state.lastFinishedMessage, suspended: state.isSuspended || undefined };
-  }
-
   // ===========================================================================
   // Control
   // ===========================================================================
@@ -3128,19 +2071,7 @@ export class Harness<TState = {}> {
    * Abort the current operation.
    */
   abort(): void {
-    // session.abortRun() drops the parked tool suspensions (so a run sitting in
-    // a tool suspend() like ask_user / request_access isn't left orphaned),
-    // aborts the live subscription, and marks the run as aborting. The Harness
-    // owns the display-state mirror of those suspensions, so clear it here too.
-    const hadPendingSuspensions = this.#session.displayState.get().pendingSuspensions.size > 0;
-    this.#session.displayState.clearPendingSuspensions();
-    this.#session.abortRun();
-    // Clearing the suspension mirror is a direct mutation, so it doesn't flow
-    // through the display-state reducer. Notify subscribers explicitly when we
-    // actually removed something, otherwise stale suspension UI can linger.
-    if (hadPendingSuspensions) {
-      this.#session.emit({ type: 'display_state_changed', displayState: this.#session.displayState.get() });
-    }
+    this.#session.abort();
   }
 
   /**
@@ -3410,7 +2341,7 @@ export class Harness<TState = {}> {
       toolsets: await this.buildToolsets(requestContext),
     });
 
-    await this.processStream(output, requestContext);
+    await this.#session.processStream(output, requestContext);
   }
 
   // ===========================================================================
